Split classes in `spec_classes.py` into separate files.
