+++ metricflow-semantics/metricflow_semantics/specs/entity_spec.py
@@ -1,0 +1,88 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, Tuple
+
+from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclass
+from dbt_semantic_interfaces.references import EntityReference
+from typing_extensions import override
+
+from metricflow_semantics.model.semantics.linkable_element import ElementPathKey, LinkableElementType
+from metricflow_semantics.naming.linkable_spec_name import StructuredLinkableSpecName
+from metricflow_semantics.specs.instance_spec import InstanceSpecVisitor, LinkableInstanceSpec
+from metricflow_semantics.visitor import VisitorOutputT
+
+
+@dataclass(frozen=True)
+class EntitySpec(LinkableInstanceSpec, SerializableDataclass):  # noqa: D101
+    @property
+    def without_first_entity_link(self) -> EntitySpec:  # noqa: D102
+        assert len(self.entity_links) > 0, f"Spec does not have any entity links: {self}"
+        return EntitySpec(element_name=self.element_name, entity_links=self.entity_links[1:])
+
+    @property
+    def without_entity_links(self) -> EntitySpec:  # noqa: D102
+        return LinklessEntitySpec.from_element_name(self.element_name)
+
+    @property
+    def as_linkless_prefix(self) -> Tuple[EntityReference, ...]:
+        """Creates tuple of linkless entities that could be included in the entity_links of another spec.
+
+        eg as a prefix to a DimensionSpec's entity links to when a join is occurring via this entity
+        """
+        return (EntityReference(element_name=self.element_name),) + self.entity_links
+
+    @staticmethod
+    def from_name(name: str) -> EntitySpec:  # noqa: D102
+        structured_name = StructuredLinkableSpecName.from_name(name)
+        return EntitySpec(
+            entity_links=tuple(EntityReference(idl) for idl in structured_name.entity_link_names),
+            element_name=structured_name.element_name,
+        )
+
+    def __eq__(self, other: Any) -> bool:  # type: ignore[misc]  # noqa: D105
+        if not isinstance(other, EntitySpec):
+            return False
+        return self.element_name == other.element_name and self.entity_links == other.entity_links
+
+    def __hash__(self) -> int:  # noqa: D105
+        return hash((self.element_name, self.entity_links))
+
+    @property
+    def reference(self) -> EntityReference:  # noqa: D102
+        return EntityReference(element_name=self.element_name)
+
+    def accept(self, visitor: InstanceSpecVisitor[VisitorOutputT]) -> VisitorOutputT:  # noqa: D102
+        return visitor.visit_entity_spec(self)
+
+    @property
+    @override
+    def element_path_key(self) -> ElementPathKey:
+        return ElementPathKey(
+            element_name=self.element_name, element_type=LinkableElementType.ENTITY, entity_links=self.entity_links
+        )
+
+
+@dataclass(frozen=True)
+class LinklessEntitySpec(EntitySpec, SerializableDataclass):
+    """Similar to EntitySpec, but requires that it doesn't have entity links."""
+
+    @staticmethod
+    def from_element_name(element_name: str) -> LinklessEntitySpec:  # noqa: D102
+        return LinklessEntitySpec(element_name=element_name, entity_links=())
+
+    def __post_init__(self) -> None:  # noqa: D105
+        if len(self.entity_links) > 0:
+            raise RuntimeError(f"{self.__class__.__name__} shouldn't have entity links. Got: {self}")
+
+    def __eq__(self, other: Any) -> bool:  # type: ignore[misc]  # noqa: D105
+        if not isinstance(other, EntitySpec):
+            return False
+        return self.element_name == other.element_name and self.entity_links == other.entity_links
+
+    def __hash__(self) -> int:  # noqa: D105
+        return hash((self.element_name, self.entity_links))
+
+    @staticmethod
+    def from_reference(entity_reference: EntityReference) -> LinklessEntitySpec:  # noqa: D102
+        return LinklessEntitySpec(element_name=entity_reference.element_name, entity_links=())
