+++ metricflow-semantics/metricflow_semantics/specs/where_filter/where_filter_spec.py
@@ -1,0 +1,80 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Tuple
+
+from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclass
+from typing_extensions import override
+
+from metricflow_semantics.collection_helpers.dedupe import ordered_dedupe
+from metricflow_semantics.collection_helpers.merger import Mergeable
+from metricflow_semantics.model.semantics.linkable_element import LinkableElement
+from metricflow_semantics.specs.instance_spec import LinkableInstanceSpec
+from metricflow_semantics.sql.sql_bind_parameters import SqlBindParameters
+
+
+@dataclass(frozen=True)
+class WhereFilterSpec(Mergeable, SerializableDataclass):
+    """Similar to the WhereFilter, but with the where_sql_template rendered and used elements extracted.
+
+    For example:
+
+    WhereFilter(where_sql_template="{{ Dimension('listing__country') }} == 'US'"))
+
+    ->
+
+    WhereFilterSpec(
+        where_sql="listing__country == 'US'",
+        bind_parameters: SqlBindParameters(),
+        linkable_specs: (
+            DimensionSpec(
+                element_name='country',
+                entity_links=('listing',),
+        ),
+        linkable_elements: (
+            LinkableDimension(
+                semantic_model_origin=SemanticModelReference(semantic_model_name='listings_latest')
+                element_name='country',
+                ...
+            )
+        )
+    )
+    """
+
+    # Debating whether where_sql / bind_parameters belongs here. where_sql may become dialect specific if we introduce
+    # quoted identifiers later.
+    where_sql: str
+    bind_parameters: SqlBindParameters
+    linkable_specs: Tuple[LinkableInstanceSpec, ...]
+    linkable_elements: Tuple[LinkableElement, ...]
+
+    def merge(self, other: WhereFilterSpec) -> WhereFilterSpec:  # noqa: D102
+        if self == WhereFilterSpec.empty_instance():
+            return other
+
+        if other == WhereFilterSpec.empty_instance():
+            return self
+
+        if self == other:
+            return self
+
+        return WhereFilterSpec(
+            where_sql=f"({self.where_sql}) AND ({other.where_sql})",
+            bind_parameters=self.bind_parameters.combine(other.bind_parameters),
+            linkable_specs=ordered_dedupe(self.linkable_specs, other.linkable_specs),
+            linkable_elements=ordered_dedupe(self.linkable_elements, other.linkable_elements),
+        )
+
+    @classmethod
+    @override
+    def empty_instance(cls) -> WhereFilterSpec:
+        # Need to revisit making WhereFilterSpec Mergeable as it's current not a collection, and it's odd to return this
+        # no-op filter. Use cases would need to check whether a WhereSpec is a no-op before rendering it to avoid an
+        # un-necessary WHERE clause. Making WhereFilterSpec map to a WhereFilterIntersection would make this more in
+        # line with other cases of Mergeable.
+        return WhereFilterSpec(
+            where_sql="TRUE",
+            bind_parameters=SqlBindParameters(),
+            linkable_specs=(),
+            linkable_elements=(),
+        )
